@@ -4,67 +4,81 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build",
-    "build:prod": "ng build --prod",
-    "build:optimized": "ng build --prod --aot --vendor-chunk --common-chunk --delete-output-path --build-optimizer",
+    "build": "ng build --stats-json",
+    "build:prod": "ng build --configuration production",
+    "build:optimized": "ng build --configuration production --aot --vendor-chunk  --common-chunk --delete-output-path --build-optimizer --stats-json",
+    "analyze": "npm run build; webpack-bundle-analyzer dist/Orso-Arpa-Web/stats.json",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.4",
-    "@angular/cdk": "~12.0.4",
-    "@angular/common": "~12.0.4",
-    "@angular/compiler": "~12.0.4",
-    "@angular/core": "~12.0.4",
-    "@angular/forms": "~12.0.4",
-    "@angular/platform-browser": "~12.0.4",
-    "@angular/platform-browser-dynamic": "~12.0.4",
-    "@angular/router": "~12.0.4",
-    "@angular/service-worker": "~12.0.2",
+    "@angular/animations": "~12.2.8",
+    "@angular/cdk": "~12.2.8",
+    "@angular/common": "~12.2.8",
+    "@angular/compiler": "~12.2.8",
+    "@angular/core": "~12.2.8",
+    "@angular/forms": "~12.2.8",
+    "@angular/material": "^12.2.8",
+    "@angular/platform-browser": "~12.2.8",
+    "@angular/platform-browser-dynamic": "~12.2.8",
+    "@angular/router": "~12.2.8",
+    "@angular/service-worker": "~12.2.8",
     "@apidevtools/json-schema-ref-parser": "^9.0.7",
+    "@apollo/client": "^3.0.0",
     "@auth0/angular-jwt": "^5.0.2",
     "@fullcalendar/core": "4.1.0",
     "@fullcalendar/daygrid": "4.1.0",
     "@fullcalendar/interaction": "4.1.0",
     "@fullcalendar/timegrid": "4.1.0",
+    "@ngrx/store": "^12.4.0",
+    "@ngrx/store-devtools": "^12.4.0",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@stoplight/json-ref-resolver": "^3.1.2",
+    "apollo-angular": "^2.6.0",
     "chart.js": "^2.9.4",
     "flag-icon-css": "^3.5.0",
+    "graphql": "^15.0.0",
     "lodash-es": "^4.17.21",
-    "ng-recaptcha": "^8.0.0",
+    "ng-recaptcha": "^8.0.1",
     "ngx-take-until-destroy": "^5.4.0",
     "primeflex": "^2.0.0",
     "primeicons": "^4.1.0",
-    "primeng": "*",
+    "primeng": "12.2.0",
+    "pwa": "^1.9.7",
     "rxjs": "~6.6.0",
     "tslib": "^2.1.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.4",
+    "@angular-builders/custom-webpack": "^12.1.1",
+    "@angular-devkit/build-angular": "~12.2.8",
     "@angular-eslint/builder": "12.2.0",
     "@angular-eslint/eslint-plugin": "12.2.0",
     "@angular-eslint/eslint-plugin-template": "12.2.0",
-    "@angular-eslint/schematics": "12.2.0",
+    "@angular-eslint/schematics": "12.5.0",
     "@angular-eslint/template-parser": "12.2.0",
-    "@angular/cli": "~12.0.4",
-    "@angular/compiler-cli": "~12.0.4",
+    "@angular/cli": "~12.2.8",
+    "@angular/compiler-cli": "~12.2.8",
+    "@ngrx/schematics": "^12.4.0",
     "@types/jasmine": "~3.6.0",
     "@types/lodash-es": "^4.17.4",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.23.0",
-    "@typescript-eslint/parser": "4.23.0",
-    "eslint": "^7.26.0",
+    "@typescript-eslint/eslint-plugin": "4.32.0",
+    "@typescript-eslint/parser": "4.32.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-graphql": "^4.0.0",
+    "eslint-plugin-ngrx": "^1.0.0",
+    "eslint-plugin-rxjs": "^3.3.7",
     "jasmine-core": "~3.7.0",
     "karma": "~6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
+    "ng2-charts": "^2.4.3",
     "typescript": "~4.2.4",
     "webpack-bundle-analyzer": "^4.4.1"
   }
