@@ -3,23 +3,41 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ConfigService } from './core/services/config.service';
-import { CoreModule } from './core/core.module';
-import { MainModule } from './main/main.module';
-import { SharedModule } from './shared/shared.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClientModule } from '@angular/common/http';
+import { MatIconModule, MatIconRegistry } from '@angular/material/icon';
+import { StoreModule } from '@ngrx/store';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { environment } from '../environments/environment';
+import { ArpaModule } from '../@arpa/arpa.module';
+import { ThemeSwitcherService } from '../@arpa/components/theme-switcher/theme-switcher.service';
+import { ConfigService } from '../@arpa/services/config.service';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import { OfflineComponent } from './components/offline/offline.component';
+import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   declarations: [
     AppComponent,
+    OfflineComponent,
   ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
     AppRoutingModule,
-    CoreModule.forRoot(),
-    MainModule,
-    SharedModule,
+    HttpClientModule,
+    StoreModule.forRoot({}, {}),
+    StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production }),
+
+    ArpaModule.forRoot(),
+    ServiceWorkerModule.register('ngsw-worker.js', {
+      enabled: environment.production,
+      // Register the ServiceWorker as soon as the app is stable
+      // or after 30 seconds (whichever comes first).
+      registrationStrategy: 'registerWhenStable:30000',
+    }),
+    TranslateModule,
+    MatIconModule,
   ],
   providers: [
     ConfigService,
@@ -33,4 +51,8 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
   bootstrap: [AppComponent],
 })
 export class AppModule {
+  constructor(private themeService: ThemeSwitcherService, private iconRegistry: MatIconRegistry) {
+    themeService.setTheme();
+    iconRegistry.registerFontClassAlias('icomoon', '');
+  }
 }
