@@ -1,15 +1,14 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { DialogService } from 'primeng/dynamicdialog';
-import { waitForAsync } from '@angular/core/testing';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { AppointmentsComponent } from './appointments.component';
 import { of } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
-import { AppointmentService } from '../../../core/services/appointment.service';
+import { AppointmentService } from '../services/appointment.service';
 import { NotificationsMockService } from '../../../../testing/notifications.mock.service';
-import { LoadingService } from '../../../core/services/loading.service';
-import { SectionService } from '../../../core/services/section.service';
+import { LoadingService } from '../../../../@arpa/services/loading.service';
+import { SectionService } from '../../../shared/services/section.service';
 import { TranslateMockModule } from '../../../../testing/translate.mock.module';
 
 describe('AppointmentsComponent', () => {
@@ -24,13 +23,13 @@ describe('AppointmentsComponent', () => {
         { provide: NotificationsMockService },
         { provide: DialogService, useValue: {} },
         { provide: LoadingService, useValue: {} },
-        { provide: SectionService, useValue: { sections$: of([])} },
-        { provide: ActivatedRoute, useValue: { data: of({})} },
+        { provide: SectionService, useValue: { sections$: of([]) } },
+        { provide: ActivatedRoute, useValue: { data: of({}) } },
       ],
       imports: [
         RouterTestingModule,
         HttpClientTestingModule,
-        TranslateMockModule
+        TranslateMockModule,
       ],
     }).compileComponents();
   }));
