@@ -2,21 +2,30 @@ import { TranslateService } from '@ngx-translate/core';
 import { ActivatedRoute } from '@angular/router';
 import { Component } from '@angular/core';
 import { SelectItem } from 'primeng/api';
-import { IAppointmentDto, ICalendarEvent, IVenueDto } from 'src/app/models/appointment';
-import { IProjectDto } from 'src/app/models/IProjectDto';
-import { DateRange } from 'src/app/models/date-range';
 import dayGridPlugin from '@fullcalendar/daygrid';
 import timeGridPlugin from '@fullcalendar/timegrid';
 import interactionPlugin from '@fullcalendar/interaction';
 import { DialogService } from 'primeng/dynamicdialog';
 import { Observable, Subscription } from 'rxjs';
 import { first, map } from 'rxjs/operators';
-import { ISectionDto } from 'src/app/models/section';
-import { AppointmentService } from '../../../core/services/appointment.service';
-import { NotificationsService } from '../../../core/services/notifications.service';
-import { SectionService } from '../../../core/services/section.service';
+import { AppointmentService } from '../services/appointment.service';
+import { NotificationsService } from '../../../../@arpa/services/notifications.service';
+import { SectionService } from '../../../shared/services/section.service';
 import { EditAppointmentComponent } from '../edit-appointment/edit-appointment.component';
-import { Unsubscribe } from '../../../core/decorators/unsubscribe.decorator';
+import { ProjectDto } from '../../../../@arpa/models/projectDto';
+import { VenueDto } from '../../../../@arpa/models/venueDto';
+import { AppointmentDto } from '../../../../@arpa/models/appointmentDto';
+import { SectionDto } from '../../../../@arpa/models/sectionDto';
+import { DateRange } from '../../../../@arpa/models/dateRange';
+import { Unsubscribe } from '../../../../@arpa/decorators/unsubscribe.decorator';
+
+export interface CalendarEvent {
+  id: string;
+  allDay: boolean;
+  start: Date;
+  end: Date;
+  title: string;
+}
 
 @Component({
   selector: 'arpa-appointments',
@@ -32,24 +41,14 @@ export class AppointmentsComponent {
   salaryPatternOptions: SelectItem[] = [];
   salaryOptions: SelectItem[] = [];
   expectationOptions: SelectItem[] = [];
-  sections: ISectionDto[] = [];
-  projects: IProjectDto[] = [];
-  venues: IVenueDto[] = [];
+  sections: SectionDto[] = [];
+  projects: ProjectDto[] = [];
+  venues: VenueDto[] = [];
   predictionOptions: SelectItem[] = [];
   resultOptions: SelectItem[] = [];
 
   fullCalendarOptions$: Observable<any>;
-  events: ICalendarEvent[] = [];
-  private _appointments: IAppointmentDto[] = [];
-
-  get appointments(): IAppointmentDto[] {
-    return this._appointments;
-  }
-
-  set appointments(values: IAppointmentDto[]) {
-    this._appointments = values;
-    this.events = values.map((a) => this.mapAppointmentToCalendarEvent(a));
-  }
+  events: CalendarEvent[] = [];
 
   constructor(
     private appointmentService: AppointmentService,
@@ -57,7 +56,7 @@ export class AppointmentsComponent {
     private route: ActivatedRoute,
     private translate: TranslateService,
     private dialogService: DialogService,
-    private sectionService: SectionService
+    private sectionService: SectionService,
   ) {
     this.route.data.pipe(first()).subscribe((data) => {
       this.projects = data.projects || [];
@@ -75,54 +74,18 @@ export class AppointmentsComponent {
     this.setOptions();
   }
 
-  private setOptions(): void {
-    this.fullCalendarOptions$ = this.translate.get('NEW').pipe(
-      map((translation) => ({
-        plugins: [dayGridPlugin, timeGridPlugin, interactionPlugin],
-        calendarWeekends: true,
-        defaultView: 'dayGridMonth',
-        defaultDate: new Date(),
-        eventResize: (e: any) => {
-          this.changeDates(e.prevEvent, e.event);
-        },
-        eventDrop: (e: any) => {
-          this.changeDates(e.oldEvent, e.event);
-        },
-        dateClick: (e: { date: Date }) => {
-          this.openCreateDialog(e.date);
-        },
-        eventClick: (e: any) => {
-          this.openEditDialog(e.event.id);
-        },
-        customButtons: {
-          btnAddAppointment: {
-            text: translation,
-            click: () => {
-              this.openCreateDialog();
-            },
-          },
-        },
-        header: {
-          left: 'dayGridMonth,timeGridWeek,timeGridDay, prevYear,prev,next,nextYear, today',
-          center: 'title',
-          right: 'btnAddAppointment',
-        },
-        editable: true,
-        locale: this.translate.currentLang,
-        firstDay: 1,
-        weekNumberCalculation: 'ISO',
-        weekNumbers: true,
-        weekNumbersWithinDays: true,
-        nowIndicator: true,
-        eventLimit: true,
-        datesRender: (info: any) => {
-          this.setAppointments(info.view.type, info.view.calendar.component.props.currentDate);
-        },
-      }))
-    );
+  private _appointments: AppointmentDto[] = [];
+
+  get appointments(): AppointmentDto[] {
+    return this._appointments;
   }
 
-  mapAppointmentToCalendarEvent(appointment: IAppointmentDto): ICalendarEvent {
+  set appointments(values: AppointmentDto[]) {
+    this._appointments = values;
+    this.events = values.map((a) => this.mapAppointmentToCalendarEvent(a));
+  }
+
+  mapAppointmentToCalendarEvent(appointment: AppointmentDto): CalendarEvent {
     const isAllDay = this.isAllDayEvent(appointment);
 
     /* fullCalendar multiple day all day events display one day short. This is an issue if an event
@@ -145,7 +108,7 @@ export class AppointmentsComponent {
     };
   }
 
-  isAllDayEvent(appointment: IAppointmentDto | undefined): boolean {
+  isAllDayEvent(appointment: AppointmentDto | undefined): boolean {
     if (appointment === undefined) {
       return false;
     }
@@ -160,7 +123,7 @@ export class AppointmentsComponent {
     return isAllDay;
   }
 
-  changeDates(oldEvent: ICalendarEvent, changedEvent: ICalendarEvent): void {
+  changeDates(oldEvent: CalendarEvent, changedEvent: CalendarEvent): void {
     let newStartTime: Date | null = null;
     let newEndTime: Date | null = null;
     if (oldEvent.start !== changedEvent.start) {
@@ -236,16 +199,63 @@ export class AppointmentsComponent {
         salaryOptions: this.salaryOptions,
         expectationOptions: this.expectationOptions,
       },
-      header: this.translate.instant('editappointments.CREATE'),
-      style: { 'max-width': '1500px' },
+      header: this.translate.instant('appointments.CREATE'),
+      styleClass: 'form-modal',
       dismissableMask: true,
     });
 
-    ref.onClose.pipe(first()).subscribe((appointment: IAppointmentDto) => {
+    ref.onClose.pipe(first()).subscribe((appointment: AppointmentDto) => {
       if (appointment) {
         this.appointments = [...this.appointments, appointment];
       }
     });
+  }
+
+  private setOptions(): void {
+    this.fullCalendarOptions$ = this.translate.get('NEW').pipe(
+      map((translation) => ({
+        plugins: [dayGridPlugin, timeGridPlugin, interactionPlugin],
+        calendarWeekends: true,
+        defaultView: 'dayGridMonth',
+        defaultDate: new Date(),
+        eventResize: (e: any) => {
+          this.changeDates(e.prevEvent, e.event);
+        },
+        eventDrop: (e: any) => {
+          this.changeDates(e.oldEvent, e.event);
+        },
+        dateClick: (e: { date: Date }) => {
+          this.openCreateDialog(e.date);
+        },
+        eventClick: (e: any) => {
+          this.openEditDialog(e.event.id);
+        },
+        customButtons: {
+          btnAddAppointment: {
+            text: translation,
+            click: () => {
+              this.openCreateDialog();
+            },
+          },
+        },
+        header: {
+          left: 'dayGridMonth,timeGridWeek,timeGridDay, prevYear,prev,next,nextYear, today',
+          center: 'title',
+          right: 'btnAddAppointment',
+        },
+        editable: true,
+        locale: this.translate.currentLang,
+        firstDay: 1,
+        weekNumberCalculation: 'ISO',
+        weekNumbers: true,
+        weekNumbersWithinDays: true,
+        nowIndicator: true,
+        eventLimit: true,
+        datesRender: (info: any) => {
+          this.setAppointments(info.view.type, info.view.calendar.component.props.currentDate);
+        },
+      })),
+    );
   }
 
   private openEditDialog(appointmentId: string): void {
@@ -265,17 +275,17 @@ export class AppointmentsComponent {
         expectationOptions: this.expectationOptions,
         isAllDayEvent: this.isAllDayEvent(appointment),
       },
-      header: this.translate.instant('editappointments.EDIT'),
-      style: { 'max-width': '1500px' },
+      header: this.translate.instant('appointments.EDIT'),
+      styleClass: 'form-modal',
       dismissableMask: true,
     });
 
-    ref.onClose.pipe(first()).subscribe((result: IAppointmentDto | string) => {
+    ref.onClose.pipe(first()).subscribe((result: AppointmentDto | string) => {
       if (result) {
         if (typeof result === 'string') {
           this.appointments.splice(
             this.appointments.findIndex((a) => a.id === appointmentId),
-            1
+            1,
           );
           this.appointments = [...this.appointments];
         } else {
