@@ -1,12 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { MuproComponent } from './mupro.component';
-import { MuproProfilesComponent } from './mupro-profiles/mupro-profiles.component';
-import { ProjectListComponent } from '../projects/project-list/project-list.component';
+import { ProfilesComponent } from './profiles/profiles.component';
 import { PersonsService } from './services/persons.service';
-import { ProjectListResolver } from '../../core/resolvers/project-list.resolver';
 import { ProfileMusicianResolver } from './resolvers/profile-musician.resolver';
 import { SectionsResolver } from '../profile/resolvers/sections.resolver';
+import { ProjectsComponent } from './projects/projects.component';
 
 const routes: Routes = [
   {
@@ -15,8 +14,9 @@ const routes: Routes = [
     resolve: { persons: PersonsService },
     children: [
       {
-        path: ':id',
-        component: MuproProfilesComponent,
+        path: ':personId',
+        component: ProfilesComponent,
+        runGuardsAndResolvers: 'always',
         resolve: {
           profiles: ProfileMusicianResolver,
           sections: SectionsResolver,
@@ -27,8 +27,12 @@ const routes: Routes = [
             redirectTo: 'projects',
           },
           {
-            path: 'projects', component: ProjectListComponent,
-            resolve: { projects: ProjectListResolver },
+            path: '',
+            outlet: 'modal',
+            loadChildren: () => import('../musician-profile-dialog/musician-profile-dialog.module').then(m => m.MusicianProfileDialogModule),
+          },
+          {
+            path: 'projects', component: ProjectsComponent,
           },
         ],
       },
