@@ -1,22 +1,18 @@
 import { Injectable } from '@angular/core';
-import {
-  Resolve,
-  RouterStateSnapshot,
-  ActivatedRouteSnapshot,
-} from '@angular/router';
+import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { PersonsService } from '../services/persons.service';
-import { IMusicianProfileDto } from '../../../models/appointment';
+import { MusicianProfileDto } from '../../../../@arpa/models/musicianProfileDto';
 
 @Injectable({
   providedIn: 'root',
 })
-export class ProfileMusicianResolver implements Resolve<IMusicianProfileDto> {
+export class ProfileMusicianResolver implements Resolve<MusicianProfileDto> {
   constructor(private personService: PersonsService) {
   }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<IMusicianProfileDto> {
-    const { id } = route.params;
-    return this.personService.getMusicianProfile(id);
+  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<MusicianProfileDto> {
+    const { personId } = route.params;
+    return this.personService.getMusicianProfile(personId);
   }
 }
