@@ -1,13 +1,14 @@
 import { Injectable } from '@angular/core';
-import { ApiService } from '../../../core/services/api.service';
+import { ApiService } from '../../../../@arpa/services/api.service';
 import { Observable } from 'rxjs';
-import { IMusicianProfileDto, IPersonDto } from '../../../models/appointment';
 import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { PersonDto } from '../../../../@arpa/models/personDto';
+import { MusicianProfileDto } from '../../../../@arpa/models/musicianProfileDto';
 
 @Injectable({
   providedIn: 'root',
 })
-export class PersonsService implements Resolve<IPersonDto[] | IPersonDto> {
+export class PersonsService implements Resolve<PersonDto[] | PersonDto> {
 
   readonly baseUrl: string;
 
@@ -15,23 +16,23 @@ export class PersonsService implements Resolve<IPersonDto[] | IPersonDto> {
     this.baseUrl = '/persons';
   }
 
-  public getPersons(): Observable<IPersonDto[]> {
-    return this.apiService.get<IPersonDto[]>(this.baseUrl);
+  public getPersons(): Observable<PersonDto[]> {
+    return this.apiService.get<PersonDto[]>(this.baseUrl);
   }
 
-  public getPerson(id: string): Observable<IPersonDto> {
-    return this.apiService.get<IPersonDto>(`${this.baseUrl}/${id}`);
+  public getPerson(id: string): Observable<PersonDto> {
+    return this.apiService.get<PersonDto>(`${this.baseUrl}/${id}`);
   }
 
-  public getMusicianProfile(id: string): Observable<IMusicianProfileDto> {
-    return this.apiService.get<IMusicianProfileDto>(`${this.baseUrl}/${id}/profiles/musician`);
+  public getMusicianProfile(id: string): Observable<MusicianProfileDto> {
+    return this.apiService.get<MusicianProfileDto>(`${this.baseUrl}/${id}/profiles/musician`);
   }
 
-  public putMusicianProfile(id: string): Observable<IMusicianProfileDto> {
-    return this.apiService.put<IMusicianProfileDto>(`${this.baseUrl}/${id}/profiles/musician`);
+  public putMusicianProfile(id: string): Observable<MusicianProfileDto> {
+    return this.apiService.put<MusicianProfileDto>(`${this.baseUrl}/${id}/profiles/musician`);
   }
 
-  public resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<IPersonDto[] | IPersonDto> {
+  public resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<PersonDto[] | PersonDto> {
     const { id } = route.data;
     return id ? this.getPerson(id) : this.getPersons();
   }
