@@ -4,39 +4,43 @@ import { ProfileComponent } from './profile.component';
 import { UserComponent } from './user/user.component';
 import { MusicianComponent } from './musician/musician.component';
 import { ProfileResolver } from './resolvers/profile.resolver';
-import { MyAppointmentsComponent } from './my-appointments/my-appointments.component';
-import { RoleNames } from '../../models/role-names';
-import { AppointmentParticipationPredictionListResolver } from '../../core/resolvers/appointment-participation-prediction-list.resolver';
+import { AppointmentsComponent } from './appointments/appointments.component';
+import { RoleNames } from '../../../@arpa/models/roleNames';
 import { ProfileMusicianResolver } from './resolvers/profile-musician.resolver';
 import { SectionsResolver } from './resolvers/sections.resolver';
+import { QRCodeComponent } from './qrcode/qrcode.component';
 
 const routes: Routes = [
   {
     path: '',
     component: ProfileComponent,
+    runGuardsAndResolvers: 'always',
     resolve: {
-      profile: ProfileResolver
+      profile: ProfileResolver,
     },
     children: [
       {
         path: '',
-        redirectTo: 'user',
-        pathMatch: 'full',
+        component: UserComponent,
+        resolve: {
+          profile: ProfileResolver,
+        },
       },
       {
         path: 'user',
         component: UserComponent,
         resolve: {
-          profile: ProfileResolver
+          profile: ProfileResolver,
         },
       },
       {
+        path: 'qrcode',
+        component: QRCodeComponent,
+      },
+      {
         path: 'appointments',
-        component: MyAppointmentsComponent,
+        component: AppointmentsComponent,
         data: { roles: [RoleNames.performer, RoleNames.staff, RoleNames.admin] },
-        resolve: {
-          predictions: AppointmentParticipationPredictionListResolver,
-        },
       },
       {
         path: 'musician',
@@ -45,9 +49,17 @@ const routes: Routes = [
           profiles: ProfileMusicianResolver,
           sections: SectionsResolver,
         },
+        runGuardsAndResolvers: 'always',
+        children: [
+          {
+            path: '',
+            outlet: 'modal',
+            loadChildren: () => import('../musician-profile-dialog/musician-profile-dialog.module').then(m => m.MusicianProfileDialogModule),
+          },
+        ],
       },
-    ]
-  }
+    ],
+  },
 ];
 
 @NgModule({
