@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ProfileService } from './profile.service';
-import { IUserProfileDto } from '../../models/IUserProfileDto';
-import { Unsubscribe } from '../../core/decorators/unsubscribe.decorator';
-import { Subscription } from 'rxjs';
+import { Observable, Subscription } from 'rxjs';
+import { MyUserProfileDto } from '../../../@arpa/models/myUserProfileDto';
+import { Unsubscribe } from '../../../@arpa/decorators/unsubscribe.decorator';
+import { map, switchMap } from 'rxjs/operators';
+import { AuthService } from '../../../@arpa/services/auth.service';
 
 @Component({
   selector: 'arpa-appointments',
@@ -14,12 +16,13 @@ import { Subscription } from 'rxjs';
 export class ProfileComponent implements OnInit {
 
   public selection: boolean;
-  public profile: IUserProfileDto;
+  public profile: Observable<MyUserProfileDto>;
   private menuEventSubscription: Subscription;
 
   constructor(
     private router: Router,
     private route: ActivatedRoute,
+    private authService: AuthService,
     private profileService: ProfileService,
   ) {
     this.menuEventSubscription = profileService.menuEvents.subscribe(() => {
@@ -28,7 +31,12 @@ export class ProfileComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.profile = this.route.snapshot.data.profile;
+    this.profile = this.authService.currentUser.pipe(switchMap(token => {
+      return this.route.data.pipe(map(({ profile }) => {
+        profile.displayName = token.displayName;
+        return profile as MyUserProfileDto;
+      }));
+    }));
   }
 
   return(event: Event) {
