@@ -1,15 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EditProjectComponent } from './edit-project.component';
-import {DynamicDialogConfig, DynamicDialogRef} from 'primeng/dynamicdialog';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {PrimeNgModule} from '../../../shared/prime-ng/prime-ng.module';
-import {RouterTestingModule} from '@angular/router/testing';
-import {HttpClientTestingModule} from '@angular/common/http/testing';
-import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
-import {httpLoaderFactory} from '../../../core/core.module';
-import {HttpClient} from '@angular/common/http';
+import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterTestingModule } from '@angular/router/testing';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { HttpClient } from '@angular/common/http';
 import { of } from 'rxjs';
+import { httpLoaderFactory } from '../../../../@arpa/translate/translate.module';
+import { DropdownModule } from 'primeng/dropdown';
+import { TableModule } from 'primeng/table';
+import { ButtonModule } from 'primeng/button';
+import { CalendarModule } from 'primeng/calendar';
+import { InputTextareaModule } from 'primeng/inputtextarea';
+import { InputTextModule } from 'primeng/inputtext';
 
 describe('EditProjectComponent', () => {
   let component: EditProjectComponent;
@@ -17,11 +22,10 @@ describe('EditProjectComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ EditProjectComponent ],
+      declarations: [EditProjectComponent],
       imports: [
         FormsModule,
         ReactiveFormsModule,
-        PrimeNgModule,
         RouterTestingModule,
         HttpClientTestingModule,
         TranslateModule.forRoot({
@@ -31,6 +35,12 @@ describe('EditProjectComponent', () => {
             deps: [HttpClient],
           },
         }),
+        DropdownModule,
+        TableModule,
+        ButtonModule,
+        CalendarModule,
+        InputTextareaModule,
+        InputTextModule,
       ],
       providers: [
         { provide: DynamicDialogRef, useValue: {} },
@@ -40,13 +50,13 @@ describe('EditProjectComponent', () => {
               projects: of([]),
               venues: of([]),
               genre: of([]),
-              state: of([])
+              state: of([]),
             },
           },
         },
       ],
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
