@@ -1,12 +1,10 @@
-import { waitForAsync } from '@angular/core/testing';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ProjectListComponent } from './project-list.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { TranslateMockModule } from '../../../../testing/translate.mock.module';
 import { ActivatedRoute } from '@angular/router';
 import { of } from 'rxjs';
-import { DummyComponent } from '../../../shared/dummy/dummy.component';
-import {DialogService} from 'primeng/dynamicdialog';
+import { DialogService } from 'primeng/dynamicdialog';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 describe('ProjectListComponent', () => {
@@ -21,10 +19,10 @@ describe('ProjectListComponent', () => {
           TranslateMockModule,
           HttpClientTestingModule,
         ],
-        declarations: [ProjectListComponent, DummyComponent],
+        declarations: [ProjectListComponent],
         providers: [
           { provide: DialogService, useValue: {} },
-          { provide: ActivatedRoute, useValue: { data: of({})} },
+          { provide: ActivatedRoute, useValue: { data: of({}) } },
         ],
       }).compileComponents();
     }),
