@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { IMusicianProfileDto } from '../../../models/appointment';
 import { Observable } from 'rxjs';
 import { concatMap, map } from 'rxjs/operators';
 import { SelectItem } from 'primeng/api';
-import { ISectionDto } from '../../../models/section';
+import { SectionDto } from '../../../../@arpa/models/sectionDto';
+import { MusicianProfileDto } from '../../../../@arpa/models/musicianProfileDto';
 
 @Component({
   selector: 'arpa-project-participation',
@@ -15,16 +15,15 @@ import { ISectionDto } from '../../../models/section';
 export class ProjectParticipationComponent implements OnInit {
 
   public form: FormGroup;
+  public sections: Observable<SectionDto> = this.config.data.sections;
   public musicianProfiles: Observable<SelectItem[]> = this.config.data.musicianProfiles.pipe(
-    concatMap((profiles: IMusicianProfileDto[]) => this.sections.pipe(
-      map((sections: any) => profiles.map((profile: IMusicianProfileDto) => {
-          const find = sections.find((section: ISectionDto) => section.id === profile.instrumentId);
-          return { label: find.name, value: profile.id} as SelectItem;
+    concatMap((profiles: MusicianProfileDto[]) => this.sections.pipe(
+        map((sections: any) => profiles.map((profile: MusicianProfileDto) => {
+          const find = sections.find((section: SectionDto) => section.id === profile.instrumentId);
+          return { label: find.name, value: profile.id } as SelectItem;
         })),
       ),
     ));
-
-  public sections: Observable<ISectionDto> = this.config.data.sections;
   public projectParticipation: Observable<any[]> = this.config.data.projectParticipation;
 
   constructor(
